[#8] 기본 layout 설정
